Fixed problem with cancelled futures
If a future was cancelled, we cannot set the result or we get an
InvalidStateException. We do not have TOCTOU here, since we're not using
threading, therefore we do not catch the exception, but rather can
conveniently check beforehand with the assurance that the future's state
has not modified until we set the value.